% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRRWdataMinN.r
\name{filterRRWdataMinN}
\alias{filterRRWdataMinN}
\title{This function uses the rrwModelList to collapse the raw data into usable RRW data}
\usage{
filterRRWdataMinN(
  data,
  grpCols = NULL,
  dataRtCol = "rt",
  dataPhitCol = "pHit",
  nCol = "n",
  dataOverlapCol = "overlapRound",
  minN = 20
)
}
\arguments{
\item{data}{This is a dataframe that is used as the input data for the RRW, which is, essentially, the data summarized by the critical condition, overlap, etc. It is output by rrwRawDataToRRWFormatData().}

\item{grpCols}{A vector of strings containing the grouping columns. If NULL this argument will be ignored.  DEFAULT = NULL}

\item{dataRtCol}{A string that identifies the name of the column in data that contains the average RTs for the specific overlap/correct/condition combination. Default is "rt"}

\item{dataPhitCol}{A string that identifies the name of the column in data that contains the proportion of trials for the specific overlap/correct/condition combination. Default is "pHit"}

\item{nCol}{A string that identifies the name of the column in data that contains the number of trials for the specific overlap/correct/condition combination. Default is "n"}

\item{dataOverlapCol}{A string that identifies the name of the column in data that contains the distributional overlaps for each row. The default is "overlapRound"}

\item{minN}{An integer that specifies the minimum number of trials required in the calculation of summary statistic in the dataRtCol and dataPhitCol for the analysis to run on that row of data. DEFAULT = 20.
'}
}
\value{
The input data for the RRW, filtered by minN
}
\description{
Function that uses the rrwModelList to collapse the raw data into usable RRW data
}
\examples{
filterRRWdataMinN (data, c("condition"), "rt", "pHit", "n", "overlapRound", 40)
}
\keyword{RRW}
\keyword{data}
\keyword{format}
\keyword{input}
