% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDRMpredValue.r
\name{getDRMpredValue}
\alias{getDRMpredValue}
\title{This function gets the predicted value of the three parameter log logisistic function returned by drc::drm().}
\usage{
getDRMpredValue(val, drmModel)
}
\arguments{
\item{val}{A number that is the "x" value for which you want the "y" value returned.}

\item{drmModel}{The DRM model returned by drc::drm().}
}
\value{
The "Y" value that corresponds to the "x" value of the "val" parameter
}
\description{
This function gets the predicted value of the three parameter log logisistic function returned by drc::drm().
}
\examples{
getDRMpredValue (drmVal, drmModel)
}
\keyword{DRM}
\keyword{log}
\keyword{logistic}
