% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupColsFromModelList.r
\name{getGroupColsFromModelList}
\alias{getGroupColsFromModelList}
\title{A function to get the grouping columns from the rrwModelList.}
\usage{
getGroupColsFromModelList(rrwModelList)
}
\arguments{
\item{rrwModelList}{A list that specifies the rrw model.  Build the rrwModelList useing rrwAddEffectToRRWModel}
}
\value{
a vector of strings that are the names of the grouping column headers.
}
\description{
This function gets the grouping columns from the rrwModelList.
}
\examples{
getGroupColsFromModelList (simpleModelList)
}
\keyword{RT}
\keyword{fit}
\keyword{morals}
\keyword{pHit}
\keyword{plots}
