% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICRRW.r
\name{getICRRW}
\alias{getICRRW}
\title{Function that calculates AIC or BIC for the entire RRW dataset}
\usage{
getICRRW(simPhit, dataPhit, simRT, dataRT, numParameters,
  equalizeRTandPhit = FALSE, ICtype = "BIC")
}
\arguments{
\item{simPhit}{a vector of numbers that is the fitted data generated from the model for the probability correct.}

\item{dataPhit}{a vector of numbers that were fitted by a model to produce simPhit.}

\item{simRT}{a vector of numbers that is the fitted data generated from the model for the RT.}

\item{numParameters}{The number of free parameters.}

\item{equalizeRTandPhit}{A boolean that specifies whether the influence of the pHit should be equal to that of rt.  Influence is a function of the number of observations in the BIC.  RT has more observations that pHit because it has both correct RTs and incorrect RTs.  If this is set to TRUE, then the rss is standardized by the number of observations. If it is set to FALSE, then the BIC is calculated as usual. DEFAULT = FALSE.}

\item{ICtype}{A string specifying whether to return the AIC or BIC.  Valid inputs are: "AIC" and "BIC".  DEFAULT = "BIC"
'}

\item{dataPhit}{a vector of numbers that were fitted by a model to produce simRT.}
}
\value{
BIC this is used as the value to be minimized by an optimization program
}
\description{
This function calculates AIC or BIC for a set of predicted values (predY) relative to actual empirical values (y).
}
\examples{
getICRRW (fittedPhit, pHit, fittedRT, rt, 5, standardize = TRUE, ICtype = "BIC")
}
\keyword{BIC}
\keyword{minimize}
