% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinLmR2.r
\name{getMinLmR2}
\alias{getMinLmR2}
\title{Function that calculates a linear model and returns (1 - r-square)}
\usage{
getMinLmR2(x, y)
}
\arguments{
\item{x}{a vector of numbers that is the predictor variable for a linear model.}

\item{y}{a vector of numbers that the outcome variable for a linear model.}
}
\value{
(1-r2) this is used as the value to be minimized by an optimization program
}
\description{
This function calculates a linear model and returns (1 - r-square).
}
\examples{
min.lm.r2 (x, y)
}
\keyword{lm}
\keyword{minimize}
\keyword{r-square}
\keyword{r2}
