% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinR2.r
\name{getMinR2}
\alias{getMinR2}
\title{Function that calculates (1 - r-square)}
\usage{
getMinR2(predY, y, na.rm = T)
}
\arguments{
\item{predY}{a vector of numbers that is the fitted data generated from a model.}

\item{y}{a vector of numbers that were fitted by a model to produce predY.}

\item{na.rm}{a boolean that specifies whether to remove NAs. DEFAULT = T.}
}
\value{
(1-r2) this is used as the value to be minimized by an optimization program
}
\description{
This function calculates (1 - r-square) for a set of predicted values (predY) relative to actual empirical values (y).
}
\examples{
getMinR2 (fittedY, y)
}
\keyword{minimize}
\keyword{r-square}
\keyword{r2}
