% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinR2RRW.r
\name{getMinR2RRW}
\alias{getMinR2RRW}
\title{Function that calculates (1 - r-square) for the entire RRW dataset}
\usage{
getMinR2RRW(
  simPhit,
  dataPhit,
  simRT,
  dataRT,
  equalizeRTandPhit = FALSE,
  na.rm = T
)
}
\arguments{
\item{simPhit}{a vector of numbers that is the fitted data generated from the model for the probability correct.}

\item{dataPhit}{a vector of numbers that were fitted by a model to produce simRT.}

\item{simRT}{a vector of numbers that is the fitted data generated from the model for the RT.}

\item{equalizeRTandPhit}{A boolean that specifies whether the influence of the pHit should be equal to that of rt.  Influence is a function of the number of observations.  RT has more observations than pHit because it has both correct RTs and incorrect RTs.  If this is set to TRUE, then the r square of the pHit and RT is calculated by equalizing the influence of the number of observations. If it is set to FALSE, then the output is calculated as usual. DEFAULT = FALSE.}

\item{na.rm}{a boolean that specifies whether to remove NAs. DEFAULT = T.}
}
\value{
(1-r2) this is used as the value to be minimized by an optimization program
}
\description{
This function calculates (1 - r-square) for a set of predicted values (predY) relative to actual empirical values (y).
}
\examples{
getMinR2RRW (fittedPhit, pHit, fittedRT, rt)
}
\keyword{minimize}
\keyword{r-square}
\keyword{r2}
