% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictedRTpHVOFit.r
\name{getPredictedRTpHVOfit}
\alias{getPredictedRTpHVOfit}
\title{A function to fit the RT and pHVO function to choice data}
\usage{
getPredictedRTpHVOfit(
  data,
  RTcol,
  pHVOcol,
  overlapRoundCol,
  useTwoParameterModel = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the summarized choice data.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each cell.}

\item{pHVOcol}{a string that specifies the name of the new column that will contains a p(HVO) for each cell.}

\item{overlapRoundCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{useTwoParameterModel}{A boolean that specifies whether to use a two parameter p(HOV) model.  If this is set to TRUE, then this function will fit a p(HVO) model whereby the rightmost point (overlap = 1.0) is not fixed at p(HVO) = 0.5. DEFAULT = FALSE.}
}
\value{
a list containing: the input data and best fit for RT and pHVO; RTfit = the lm object; pHitFit = the nls fit object with the fit of the pHVO data.
}
\description{
This function attempts to fit a non-linear, decelerating function to the pHVO data; a linear function to the RT data.
}
\examples{
getPredictedRTpHVOfit (data=moralsData,"resdRT", "pHit", "overlapRound", useTwoParameterModel = FALSE)
}
\keyword{RT}
\keyword{fit}
\keyword{pHit}
\keyword{plots}
