% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictedRTpHVOfitByGroup.r
\name{getPredictedRTpHVOfitByGroup}
\alias{getPredictedRTpHVOfitByGroup}
\title{A function to fit the RT and pHVO function to choice data for different groups}
\usage{
getPredictedRTpHVOfitByGroup(
  data,
  grpCols = NULL,
  RTcol,
  overlapRoundCol,
  minNperOverlap = 0,
  useTwoParameterModel = TRUE
)
}
\arguments{
\item{data}{A dataframe containing the choice data (often after running through ch.moralsDataPrep()).}

\item{grpCols}{A vector of strings containing the grouping columns.  If NULL, then the fit will be run without grouping. DEFAULT = NULL.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}

\item{overlapRoundCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{minNperOverlap}{an integer that specifies the minimum number of trials necessary to include an overlap bin in the graph. DEFAULT = 0.}

\item{useTwoParameterModel}{A boolean that specifies whether to use a two parameter p(HOV) model.  If this is set to TRUE, then this function will fit a p(HVO) model whereby the rightmost point (overlap = 1.0) is not fixed at p(HVO) = 0.5. DEFAULT = TRUE.}
}
\value{
a list containing: the input data and best fit for RT and pHVO; RTfit = the lm object for each group; pHitFit = the nls fit object with the fit of the pHit data for each group.
}
\description{
This function attempts to fit a non-linear, decelerating function to the pHVO data; a linear function to the RT data.
}
\examples{
getPredictedRTpHVOfitForModelList (data=moralsData,simpleModelList, "resdRT", "overlapRound", minNperOverlap = 40, useTwoParameterModel = TRUE)
}
\keyword{RT}
\keyword{fit}
\keyword{pHit}
\keyword{plots}
