% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictedRTpHVOFit.old.r
\name{getPredictedRTpHVOfitOld}
\alias{getPredictedRTpHVOfitOld}
\title{A function to fit the RT and pHVO function to choice data}
\usage{
getPredictedRTpHVOfitOld(
  data,
  RTcol,
  pHVOcol,
  overlapRoundCol,
  grpCol = NULL,
  useTwoParameterModel = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the summarized choice data.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each cell.}

\item{pHVOcol}{a string that specifies the name of the new column that will contains a p(HVO) for each cell.}

\item{overlapRoundCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{grpCol}{a string that specifies the name of the column in "data" that contains the grouping variable that identifies those stimuli whose value are above the reference distribution and those that are below the reference distribution. When this variable is included and useTwoParameterModel is set to TRUE, then the alpha parameter of the p(HVO) function will be fit so it is symetric above and below 0.5 for refHVO and refLVO. Also, the intercept parameter of the RT function is allowed to vary for refHVO and refLVO, but the slope is kept constant. DEFAULT = NULL (the grouping variable is ignored)}

\item{useTwoParameterModel}{A boolean that specifies whether to use a two parameter p(HOV) model.  If this is set to TRUE, then this function will fit a p(HVO) model whereby the rightmost point (overlap = 1.0) is not fixed at p(HVO) = 0.5. DEFAULT = FALSE.}
}
\value{
a list containing: the input data and best fit for RT and pHVO; RTfit = the lm object; pHitFit = the nls fit object with the fit of the pHVO data.
}
\description{
This function attempts to fit a non-linear, decelerating function to the pHVO data; a linear function to the RT data.
}
\examples{
getPredictedRTpHVOfit (data=moralsData,"resdRT", "pHit", "overlapRound", grpCol = "refValue", useTwoParameterModel = FALSE)
}
\keyword{RT}
\keyword{fit}
\keyword{pHit}
\keyword{plots}
