% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.r2.r
\name{min.r2}
\alias{min.r2}
\title{Function that calculates (1 - r-square)}
\usage{
\method{min}{r2}(predY, y)
}
\arguments{
\item{predY}{a vector of numbers that is the fitted data generated from a model.}

\item{y}{a vector of numbers that were fitted by a model to produce predY.}
}
\value{
(1-r2) this is used as the value to be minimized by an optimization program
}
\description{
This function calculates (1 - r-square) for a set of predicted values (predY) relative to actual empirical values (y).
}
\examples{
min.r2 (fittedY, y)
}
\keyword{minimize}
\keyword{r-square}
\keyword{r2}
