% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRRWfit2.r
\name{plotRRWFit2}
\alias{plotRRWFit2}
\title{This function plots the fitted RRW simulation as a function of the empirical data.}
\usage{
plotRRWFit2(
  data,
  dataRtCol = "rt",
  dataPhitCol = "pHit",
  rtFitCol = "rtFit",
  pHitFitCol = "pCross",
  correctCol = "correct",
  overlapCol = "overlap",
  condCol = NULL,
  numSimsToPlot = 0,
  plotFilename = NULL,
  multiplePlotsPerPage = TRUE,
  yMinMixRT = NULL,
  xMinMax = NULL,
  combineRThvoRTlvo = FALSE
)
}
\arguments{
\item{data}{This is a dataframe that must contain the following columns: overlap; RT (often a median); the proportion correct/incorrect; whether or not the row specifies a correct or incorrect trial. The dataset can also contains columns that effect code the influence of different parameters.}

\item{dataRtCol}{A string that identifies the name of the column in data that contains the RTs for the specific overlap/correct/condition combination. Default is "rt"}

\item{dataPhitCol}{A string that identifies the name of the column in data that contains the proportion of trials that are either correct or incorrect for the specific overlap/correct/condition combination. Default is "pHit"}

\item{rtFitCol}{A string that identifies the name of the column in data that contains the RRW fit of the RTs for the specific overlap/correct/condition combination. Default is "rtFit"}

\item{pHitFitCol}{A string that identifies the name of the column in data that contains the RRW fit of the proportion of trials that are either correct or incorrect for the specific overlap/correct/condition combination. Default is "pCross"}

\item{correctCol}{A string that identifies the name of the column in data that identifies whether the trials were correct (TRUE) or incorrect (FALSE). The default is "correct"}

\item{overlapCol}{A string that identifies the name of the column in data that contains the distributional overlaps for each row. The default is "overlap"}

\item{condCol}{A string that identifies the name of the column in data that identifies the conditions that will be plotted as separate lines. DEFAULT = NULL}

\item{numSimsToPlot}{A number specifying how many simulation runs are in the dataset and should be plotted. Default is 0, indicating that no simulations should be plotted.}

\item{plotFilename}{A string that identifies the name of file (.pdf) in which the plot will be saved. The default is NULL, whereby the plot will not be saved.}

\item{multiplePlotsPerPage}{A boolean that identifies whether to print multiple plots per page.  DEFAULT = TRUE.}

\item{yMinMixRT}{A vector of 2 numbers that identifies the c(min, max) for the y-axis of the RT graph. If not entered, then the function will calculate a pretty min and max. DEFAULT = NULL.}

\item{xMinMixRT}{A vector of 2 numbers that identifies the c(min, max) for the x-axis (overlap). DEFAULT = c(0,1).}
}
\value{
.
}
\description{
Function plots the fitted RRW simulation as a function of the empirical data.  It is generally run after the getRRWfit().
}
\examples{
plotRRWFit (data, "rt", "pHit", "rtFit", "pHitFit", "correct", "overlap")
}
\keyword{RRW}
\keyword{fit}
\keyword{output}
\keyword{plot}
\keyword{random}
\keyword{walk}
