% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrwAddParameterEffectListToRRWModel.r
\name{rrwAddParameterEffectListToRRWModel}
\alias{rrwAddParameterEffectListToRRWModel}
\title{This  function is used to create or modify an rrwModelList}
\usage{
rrwAddParameterEffectListToRRWModel(
  rrwModelList = NULL,
  parameterEffectList = NULL
)
}
\arguments{
\item{rrwModelList}{A list that specifies the rrw model.  If you are building a new list, then this should be NULL. If you are adding a new effect to an old list, then this should be the rrwModelList that you are adding an effect to. DEFAULT = NULL}

\item{parameterEffectList}{A list that specifies the parameter effect (or a list of effects, e.g., c(effect1, effect2, effect3)).  DEFAULT = NULL
'}
}
\value{
the updated rrwModelList
}
\description{
Function that creates or modifys an rrwModelList by adding effects to the list one at a time.
}
\examples{
rrwAddParameterEffectListToRRWModel (modelList, parameterEffectList)
}
\keyword{add}
\keyword{effect}
\keyword{parameter}
\keyword{rrw}
\keyword{rrwModelList}
\keyword{update}
