% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrwConvertFreeRRWModelToFixedRRWModel.r
\name{rrwConvertFreeRRWModelToFixedRRWModel}
\alias{rrwConvertFreeRRWModelToFixedRRWModel}
\title{This function modifies an rrwModelList so that the parameters values are fixed to the output of the parameter values in a runStats list (which is the output of the rrw fit functions). It is used to test a fixed parameter model on data using the model parameters extracted from another dataset.}
\usage{
rrwConvertFreeRRWModelToFixedRRWModel(rrwModelList, runStatsList)
}
\arguments{
\item{rrwModelList}{A list that specifies the rrw model.  Build the rrwModelList useing rrwAddEffectToRRWModel}

\item{runStats}{A list that contains the output of a single run of the kFold cross validation of the RRW.  It is output by rrw fit functions and used in the rrwKfoldStatsToDataframe, etc.
'}
}
\value{
a rrwModelList that has the parameters values are fixed to the output of the parameter values in a runStats list.
}
\description{
Function that modifies the parameter values of an rrwModelList to those of a runStats list.
}
\examples{
rrwConvertFreeRRWModelToFixedRRWModel (rrwModelList,runStats)
}
\keyword{RRW}
\keyword{modelList}
\keyword{rrw}
\keyword{runStats}
