% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrwCreateParameterEffect.r
\name{rrwCreateParameterEffect}
\alias{rrwCreateParameterEffect}
\title{This  function is used to creates am rrwParemeterEffect for an rrwModelList}
\usage{
rrwCreateParameterEffect(
  parameter = NULL,
  columnName = NULL,
  ParameterName = NULL,
  parameterBounds = c(0, 0, 0.001),
  df.code = NULL,
  GroupByVariables = NULL
)
}
\arguments{
\item{parameter}{A string that specifies the parameter that the effect being added will influence.  It can take on one of the following values: "s","b","nSD", "db". "da". "vc" DEFAULT = NULL}

\item{columnName}{A string that specifies the name of the column that will be created and added to the data.  The column will contain the effect/dummy code of this effect. This columnName can remain NULL if this effect is constant across all conditions (therefore the dummycode would equal 1 for all conditions and therefore would be redundant). DEFAULT = NULL}

\item{ParameterName}{A string that specifies the name of the free parameter for this effect.  It is not the "parameter" argument above.  It is, rather, a variable name that holds the free parameter values and is output in the output file. You will identify this influence of this effect by the ParameterName. DEFAULT = NULL.}

\item{parameterBounds}{A vector that specifies the upperBound, lowerBound, and minimum interval, c(upper, lower, interval), that the free parameter will vary when optimized. If the upperBound==lowerBound, then this will be a fixed parameter, that is fixed at the upperBound value. DEFAULT = c(0,0,.001)}

\item{df.code}{A dataframe that specifies the conditional statements for coding the dummy or effect variable.  The first column is a string stating the conditional (e.g., "variableName == 'level'"), the second column is a numeric stating the value (1). Each row, is a different conditional/value. The last row, first column, must be "default" and the second column must be the default value.  This df.code can remain NULL if this effect is constant across all conditions (therefore there are no conditionals). DEFAULT = NULL.}

\item{GroupByVariables}{A vector of strings that specify the variable names of the grouping variable(s) that will be used by the ddply to create the summary dataset.  DEFAULT = NULL.
'}
}
\value{
am rrwModelList that only contains the parameter effect input
}
\description{
Function that creates am rrwParemeterEffect for an rrwModelList.
}
\examples{
rrwCreateParameterEffect (parameter = "b", columnName = NULL, ParameterName = overallB, parameterBounds = c(100,10,0.001), df.code = NULL, GroupByVariables = NULL))
}
\keyword{add}
\keyword{create}
\keyword{effect}
\keyword{parameter}
\keyword{rrw}
\keyword{rrwModelList}
