% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrwRawDataToRRWFormatData.r
\name{rrwRawDataToRRWFormatData}
\alias{rrwRawDataToRRWFormatData}
\title{This function uses the rrwModelList to collapse the raw data into usable RRW data}
\usage{
rrwRawDataToRRWFormatData(data, rrwModelList, dataRtCol = "rt",
  correctCol = "correct", correctVals = c(TRUE, FALSE),
  dataOverlapCol = "overlapRound")
}
\arguments{
\item{data}{This is a dataframe containing the raw trial-by-trial data that must contain the following columns: overlap; RT (often a median); the proportion correct/incorrect; whether or not the row specifies a correct or incorrect trial. The dataset can also contains columns that effect code the influence of different parameters.}

\item{rrwModelList}{A list that specifies the rrw model.  Build the rrwModelList useing rrwAddEffectToRRWModel}

\item{dataRtCol}{A string that identifies the name of the column in data that contains the RTs for the specific overlap/correct/condition combination. Default is "rt"}

\item{correctCol}{a string that specifies the name of the column that specifies if the participant chose the item with the greatest value distribution (correct) or if they did not (incorrect).}

\item{correctVals}{a vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no"). DEFAULT = c(TRUE, FALSE)}

\item{dataOverlapCol}{A string that identifies the name of the column in data that contains the distributional overlaps for each row. The default is "overlap"
'}
}
\value{
The input data for the RRW, which is, essentially, the data summarized by the critical condition, overlap, etc.
}
\description{
Function that uses the rrwModelList to collapse the raw data into usable RRW data
}
\examples{
rrwRawDataToRRWFormatData (data, modelList, "rt", "correct", c(TRUE, FALSE), "overlapRound")
}
\keyword{RRW}
\keyword{data}
\keyword{format}
\keyword{input}
