% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeDataAndFit.r
\name{standardizeDataAndFit}
\alias{standardizeDataAndFit}
\title{Function that standardizes the data and fit of the simulation}
\usage{
standardizeDataAndFit(data, fit)
}
\arguments{
\item{data}{a vector of numbers that are the behavioral data.}

\item{fit}{a vector of numbers that were fit to the data by a model.}
}
\value{
a dataframe with the standardized data and fit
}
\description{
This function standardizes the data and fit of the simulation
}
\examples{
standardizeDataAndFit (data, fit)
}
\keyword{scale}
\keyword{standardize}
